src/requirements.txt
 mock-boto3==0.0.5
 pytest==8.2.1
 boto3==1.34.113
 botocore==1.34.113
psycopg2-binary==2.9.9